Move get-site-api to utils
